print error and sample counts to txt file

+++ include/ApolloSM/eyescan.hh
@@ -7,19 +7,19 @@ struct eyescanCoords {
   int voltage;
   double phase;
   double BER;
-  int sample0;
-  int error0;
-  int sample1;
-  int error1;
+  unsigned long int sample0;
+  unsigned long int error0;
+  unsigned long int sample1;
+  unsigned long int error1;
   uint8_t voltageReg;
   uint16_t phaseReg;
 };
 struct SESout{ //single eyescan output
 	double BER;
-	int sample0;
-	int error0;
-	int sample1;
-	int error1;
+	unsigned long int sample0;
+	unsigned long int error0;
+	unsigned long int sample1;
+	unsigned long int error1;
 };
 
 #endif
